refac(91026): Extrai task gerar_arquivo_ata_async
